+++ package.json
@@ -13,6 +13,7 @@
     "@infisical/sdk": "^2.2.3",
     "@nextui-org/react": "^2.4.1",
     "@shopify/storefront-api-client": "^1.0.0",
+    "@tanstack/react-query": "^5.45.1",
     "@xmtp/frames-validator": "^0.6.1",
     "framer-motion": "^11.2.10",
     "frames.js": "^0.16.5",
